Friday phrase randomizer
Friday task loop can now call friday_phrase_randomizer() functon to return a string to add to the friday task loop reply. We can add as many phrases as we want in the Friday/phrases.json file following the format "#" : "silly phrase". I didn't add this file to the Data folder since that is in the gitignore.