Delete func memory using threading.
Delete function has been reworked to allow the bot to suggest titles to delete if there are multiple instances of a game with the same name (ex. delete "Lego" could be asking to delete any number of the top tier lego star wars games). This functions very similarly to how add games works if the user does not input an exact title name.

There has also been a small retouching in terms of formatting. The bot will now tell you the the current price of a game and if it is on sale when you add it to your wishlist. Some other small changes like changing wording from "our tracking list" to "your wishlist" to make the bot seem more personable and it is the user in control.